refactor command executor
